refactor(proxy): reorganize proxy utilities and introduce new header processing functions
